+++ src/fork_core/application/mod.rs
@@ -5,3 +5,4 @@
 
 pub mod inbound_message;
 pub mod services;
+pub mod use_cases;